fix(home): address tabAdapter memory leak
