+++ app/src/main/java/com/louis/app/cavity/ui/home/widget/ScrollableTabAdapter.kt
@@ -3,10 +3,8 @@ package com.louis.app.cavity.ui.home.widget
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.NO_ID
 import com.louis.app.cavity.R
 
@@ -14,14 +12,14 @@ class ScrollableTabAdapter<T>(
     private val onTabClick: (View, Int) -> Unit,
     private val onLongTabClick: (T, Int) -> Unit
 ) :
-    ListAdapter<T, ScrollableTabAdapter<T>.TabViewHolder>(ScrollableItemDiffCallback<T>()) {
+    ListAdapter<T, TabViewHolder<T>>(ScrollableItemDiffCallback<T>()) {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TabViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TabViewHolder<T> {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.item_county, parent, false)
-        return TabViewHolder(view)
+        return TabViewHolder(view, onTabClick, onLongTabClick)
     }
 
-    override fun onBindViewHolder(holder: TabViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: TabViewHolder<T>, position: Int) {
         holder.bind(getItem(position))
     }
 
@@ -42,25 +40,6 @@ class ScrollableTabAdapter<T>(
             oldItem.toString() == newItem.toString()
 
         override fun areContentsTheSame(oldItem: T & Any, newItem: T & Any) = true
-    }
-
-    inner class TabViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
-        private val textView: TextView = view.findViewById(R.id.county)
-
-        init {
-            view.setOnClickListener {
-                onTabClick(it, bindingAdapterPosition)
-            }
-        }
-
-        fun bind(item: T) {
-            textView.text = item.toString()
-
-            view.setOnLongClickListener {
-                onLongTabClick(item, bindingAdapterPosition)
-                true
-            }
-        }
     }
 }
